update with from-new-20060628-1

+++ decode.c
@@ -77,7 +77,7 @@ char *decode(char *p)
 		    *q++ = ch;
 	    }
 	    p += 2;
-	    while (isspace(*p))
+	    while (isspace((int)*p))
 		p++;
 	} else {		/* no "=?ISO-2022-JP?B?" */
 	    *q++ = *p++;
