+++ from.c
@@ -49,7 +49,7 @@ int from(POP_SESSION *psp, int msg, int width)
 	if (*p == NUL || *p == '\n')	/* null line */
 	    continue;
 	if (header) {
-	    if (isspace(*p) && state) {
+	    if (isspace((int)*p) && state) {
 		strcat(heading[state], "\n");
 		strcat(heading[state], p);
 	    } else if (strncasecmp(p, "from:", 5) == 0)
